+++ lib/calendar/calendar.php
@@ -70,6 +70,7 @@ class calendar extends service
 		foreach ($events as $event) {
 			if (!isset($calmetadata['calendardesc'])) $calmetadata['calendardesc'] = "";
 			if (!isset($calmetadata['calendarcolor'])) $calmetadata['calendarcolor'] = "";
+			
 			$this->addData(array(
 				'start' => $ical->iCalDateToUnixTimestamp($event->dtstart, true),
 				'end' => $event->dtend != null ? $ical->iCalDateToUnixTimestamp($event->dtend, true) : $ical->iCalDateToUnixTimestamp($event->dtstart, true),
@@ -78,7 +79,8 @@ class calendar extends service
 				'where' => $event->location,
 				'calendarname' => $calmetadata['calendarname'] != '' ? $calmetadata['calendarname'] : $ical->calendarName(),
 				'calendardesc' => $calmetadata['calendardesc'] != '' ? $calmetadata['calendardesc'] : $ical->calendarDescription(),
-				'calendarcolor' => $calmetadata['calendarcolor'] != '' ? $calmetadata['calendarcolor'] : ""
+				'calendarcolor' => $calmetadata['calendarcolor'] != '' ? $calmetadata['calendarcolor'] : "",
+				'class' => isset($event->class) ? $event->class : ''
 				//,'link' => ''
 			));
 		}