+++ changelog.md
@@ -6,6 +6,7 @@
 - status.collapse now has an option to define the action (hide / show) performed when the trigger item reaches a specified value
 - basic.input: new modes datetime and datetimeflip allow setting datetime items with date and time in one step
 - basic.window & design.window provide more styling capabilities
+- calendar.list changes event title to "private appointment" if event is marked as private. Enable with new entry 'calendar_private = "1" ' in config.ini (under development)
 
 ### Other New Features
 - if the configured driver is not available (e.g. removed after deprecation) a warning is displayed and the offline driver is used (also in config page) in order to throw no errors due to missing io.